requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
catboost>=1.2.0
prometheus-client>=0.20.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
pytest>=8.0.0
httpx>=0.27.0
